Fe room ui

See merge request s05-webmobile1-sub3/S05P13C104!3

+++ backend/src/main/java/com/ssafy/api/response/ConferenceRes.java
@@ -13,9 +13,9 @@ import java.util.List;
 @ApiModel("ConferenceResponse")
 public class ConferenceRes extends BaseResponseBody {
     @ApiModelProperty(name="")
-    private Long roomId;
+    private int roomId;
 
-    public static ConferenceRes of(Integer statusCode, String message, Long roomId) {
+    public static ConferenceRes of(Integer statusCode, String message, int roomId) {
         ConferenceRes res = new ConferenceRes();
         res.setStatusCode(statusCode);
         res.setMessage(message);
